# Conflicts:
#	.gradle/6.3/javaCompile/classAnalysis.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.idea/misc.xml
#	build/classes/java/main/com/hotel/controller/CartController.class
#	build/classes/java/main/com/hotel/controller/CheckoutController.class
#	build/classes/java/main/com/hotel/database/item/CartDatabase.class
#	build/classes/java/main/com/hotel/database/item/CartItemDatabase.class
#	build/classes/java/main/com/hotel/database/item/MenuItemDatabase.class
#	build/classes/java/main/com/hotel/database/item/OrderDatabase.class
#	build/classes/java/main/com/hotel/database/item/intf/CartDatabaseInterface.class
#	build/classes/java/main/com/hotel/database/item/intf/CartItemDatabaseInterface.class
#	build/classes/java/main/com/hotel/database/item/intf/MenuItemDatabaseInterface.class
#	build/classes/java/main/com/hotel/database/item/intf/OrderDatabaseInterface.class
#	build/classes/java/main/com/hotel/database/jpa/JpaCartItemRepository.class
#	build/classes/java/main/com/hotel/database/jpa/JpaCartRepository.class
#	build/classes/java/main/com/hotel/database/jpa/JpaMenuItemRepository.class
#	build/classes/java/main/com/hotel/database/jpa/JpaOrderRepository.class
#	build/classes/java/main/com/hotel/database/jpa/JpaRequestRepository.class
#	build/classes/java/main/com/hotel/database/load/DatabaseLoader.class
#	build/classes/java/main/com/hotel/database/request/AbstractRequestDatabase.class
#	build/classes/java/main/com/hotel/database/request/RequestDatabase.class
#	build/classes/java/main/com/hotel/database/request/RequestWithNotesDatabase.class
#	build/classes/java/main/com/hotel/database/request/WakeUpRequestDatabase.class
#	build/classes/java/main/com/hotel/database/request/intf/AbstractRequestDatabaseInterface.class
#	build/classes/java/main/com/hotel/database/request/intf/RequestWithNotesDatabaseInterface.class
#	build/classes/java/main/com/hotel/database/request/intf/WakeUpRequestDatabaseInterface.class
#	build/classes/java/main/com/hotel/model/item/Cart.class
#	build/classes/java/main/com/hotel/model/item/CartItem.class
#	build/classes/java/main/com/hotel/model/item/MenuItem.class
#	build/classes/java/main/com/hotel/model/item/Order.class
#	build/classes/java/main/com/hotel/service/item/CartItemService.class
#	build/classes/java/main/com/hotel/service/item/CartService.class
#	build/classes/java/main/com/hotel/service/item/OrderService.class
#	build/classes/java/main/com/hotel/service/item/intf/CartItemServiceInterface.class
#	build/classes/java/main/com/hotel/service/item/intf/CartServiceInterface.class
#	build/classes/java/main/com/hotel/service/item/intf/MenuItemServiceInterface.class
#	build/classes/java/main/com/hotel/service/item/intf/OrderServiceInterface.class
#	build/resources/main/templates/cart.html
#	data/db.mv.db
#	src/main/java/com/hotel/controller/CartController.java
#	src/main/resources/templates/cart.html

+++ src/main/java/com/hotel/database/item/MenuItemDatabase.java
@@ -55,7 +55,7 @@ public class MenuItemDatabase implements MenuItemDatabaseInterface {
      * @param id The id of the item to remove from the database
      */
     @Override
-    public void remove(int id) {
+    public void remove(long id) {
         repo.deleteById(id);
     }
 
@@ -68,7 +68,7 @@ public class MenuItemDatabase implements MenuItemDatabaseInterface {
      * @param new_description The item's new description
      */
     @Override
-    public void edit(@NotNull int id, String new_name, Set<String> new_allergens, double new_price, String new_description) {
+    public void edit(@NotNull long id, String new_name, Set<String> new_allergens, double new_price, String new_description) {
 
         MenuItem item = find(id);
 
@@ -91,7 +91,7 @@ public class MenuItemDatabase implements MenuItemDatabaseInterface {
      * @return the item with the given id, or null if no such item exists
      */
     @Override
-    public MenuItem find(int id) {
+    public MenuItem find(long id) {
 
         Optional<MenuItem> item = repo.findById(id);
 
@@ -121,7 +121,7 @@ public class MenuItemDatabase implements MenuItemDatabaseInterface {
      * @return true if the id exists in the database
      */
     @Override
-    public boolean containsId(int id) {
+    public boolean containsId(long id) {
         return repo.existsById(id);
     }
 }
