+++ src/main/java/com/hotel/database/item/intf/OrderDatabaseInterface.java
@@ -12,14 +12,14 @@ public interface OrderDatabaseInterface {
 
     void add(Order order);
 
-    void edit(@NotNull int id, int new_cart_id, String new_status, long new_time) throws Exception;
+    void edit(@NotNull long id, long new_cart_id, String new_status, long new_time) throws Exception;
 
     void remove(Order order);
 
-    void remove(int id);
+    void remove(long id);
 
-    Order find(int id);
+    Order find(long id);
 
-    boolean contains(int id);
+    boolean contains(long id);
 
 }
