+++ src/main/java/com/hotel/database/request/intf/AbstractRequestDatabaseInterface.java
@@ -12,11 +12,11 @@ public interface AbstractRequestDatabaseInterface {
 
     void remove(AbstractRequest request);
 
-    void remove(int ID);
+    void remove(long ID);
 
-    AbstractRequest find(int ID);
+    AbstractRequest find(long ID);
 
     ArrayList<AbstractRequest> findRoom(int room);
 
-    boolean containsID(int ID);
+    boolean containsID(long ID);
 }
