+++ src/main/java/com/hotel/service/request/GeneralRequestService.java
@@ -20,7 +20,7 @@ public class GeneralRequestService extends AbstractRequestService implements Req
     }
 
     @Override
-    public void edit(int ID, int new_room, long new_time, String new_notes, String new_commonRequestType) {
+    public void edit(long ID, int new_room, long new_time, String new_notes, String new_commonRequestType) {
         database.edit(ID, new_room, new_time, new_notes, new_commonRequestType);
     }
 
@@ -45,12 +45,12 @@ public class GeneralRequestService extends AbstractRequestService implements Req
     }
 
     @Override
-    public void remove(int ID) {
+    public void remove(long ID) {
         database.remove(ID);
     }
 
     @Override
-    public AbstractRequest find(int ID) {
+    public AbstractRequest find(long ID) {
         return database.find(ID);
     }
 
@@ -60,7 +60,7 @@ public class GeneralRequestService extends AbstractRequestService implements Req
     }
 
     @Override
-    public boolean exists(int ID) {
+    public boolean exists(long ID) {
         return database.containsID(ID);
     }
