+++ src/main/java/com/hotel/service/request/intf/RequestServiceInterface.java
@@ -4,6 +4,6 @@ import com.hotel.model.request.Request;
 
 public interface RequestServiceInterface extends AbstractRequestServiceInterface {
 
-    void edit(int ID, int new_room, long new_time, Request.RequestType new_type);
+    void edit(long ID, int new_room, long new_time, Request.RequestType new_type);
 
 }
