+++ src/main/java/com/hotel/service/request/intf/RequestWithNotesServiceInterface.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 
 public interface RequestWithNotesServiceInterface extends AbstractRequestServiceInterface {
 
-    void edit(int ID, int new_room, long new_time, String new_notes, String new_commonRequestType);
+    void edit(long ID, int new_room, long new_time, String new_notes, String new_commonRequestType);
 
     ArrayList<RequestWithNotes> findByCommonRequestType(String commonRequestType);
